extract logic from `instrumentation.ts` to `registerSQLFunctions`
