+++ app/src/app/api/private/refresh/route.ts
@@ -1,6 +1,6 @@
 import { analysis } from '@/src/analysis/analysis'
 import { isProd } from '@/src/app/_components/utils/isEnv'
-import { register } from '@/src/instrumentation'
+import { registerSQLFunctions } from '@/src/instrumentation/registerSQLFunctions'
 import { NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 import { guardEnpoint } from '../guardEndpoint'
@@ -13,7 +13,7 @@ export async function GET(req: NextRequest) {
   const { access, response } = guardEnpoint(req, Schema)
   if (!access) return response
   try {
-    await register()
+    await registerSQLFunctions()
     analysis()
     return NextResponse.json({ message: 'OK' }, { status: 200 })
   } catch (e) {
